AI_2/requirements.txt
anytree==2.8.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.0
cycler==0.11.0
dill==0.3.6
ffmpeg-python==0.2.0
fonttools==4.37.1
future==0.18.3
IBMQuantumExperience==2.0.4
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
mlxtend==0.21.0
mpmath==1.2.1
musdb==0.4.0
ntlm-auth==1.5.0
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pbr==5.11.1
Pillow==9.2.0
ply==3.11
psutil==5.9.4
pyaml==21.10.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
qiskit==0.39.5
qiskit-aer==0.11.2
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.22.4
requests==2.28.2
requests-ntlm==1.1.0
retworkx==0.12.1
rustworkx==0.12.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
six==1.16.0
sklearn==0.0
stempeg==0.2.3
stevedore==4.1.1
sympy==1.11.1
tabulate==0.8.10
threadpoolctl==3.1.0
tqdm==4.65.0
tweedledum==1.1.1
urllib3==1.26.14
websocket-client==1.4.2
websockets==10.4
youtube-dl==2021.12.17
